% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertAndCorrect.R
\name{convert}
\alias{convert}
\title{A function converting a portfolio-stock dataframe to a stock-count matrix}
\usage{
convert(d)
}
\arguments{
\item{d}{data frame with each row containing the first 10 core stocks of a portfolio}
}
\value{
a list containing the converted matrix and whether there are false data
}
\description{
A function converting a portfolio-stock dataframe to a stock-count matrix
}
