% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertAndCorrect.R
\name{correct.error}
\alias{correct.error}
\title{correct errors from the result of the function convert}
\usage{
correct.error(list)
}
\arguments{
\item{list}{result of the function convert}
}
\value{
return a correct matrix
}
\description{
if there are false data in the result it, find it and correct it
}
