% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
A gibbs sampler using Rcpp
}
