% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertAndCorrect.R
\name{lmlv19}
\alias{lmlv19}
\title{A dataset used to describe how to use the function convert}
\description{
This dataset contains no false data, so it can be directly converted by function convert
}
\examples{
\dontrun{
data(lmlv19)
attach(lmlv19)
l <- convert(lmlv19)
str(l)
}
}
