\name{SA23204171-package}
\alias{SA23204171-package}
\docType{package}
\title{
  convert and correct portfolio-stock data
}
\description{
  Given a data frame of which each row containing the first 10 core stocks of a 
  portfolio, convert it to a matrix with its rownames being the names of all 
  core stocks in the data frame, and the rest of rows marking whether this core 
  stock appears in the corresponding portfolio with 1 and 0. Meanwhile there 
  might be false data in some portfolios, and we find them and correct them.
}
\author{
SA23204171

Maintainer: SA23204171 <sy200149@mail.ustc.edu.cn>
}
\keyword{ package }
