% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertAndCorrect.R
\name{smsv19}
\alias{smsv19}
\title{A dataset used to describe how to use the function correct.error}
\description{
This dataset contains errors, so it can be corrected by function correct.error
}
\examples{
\dontrun{
data(smsv19)
attach(smsv19)
l <- convert(smsv19)
res <- correct.error(l)
str(res)
}
}
